Move pagination helper into helpers package after middleware restructure
Merge dev's response->helpers rename and update the pagination helper to
match (WritePaginated using WriteJSON) instead of the stale response/utils
split.
